// frontend/generated/models/ProjectCategoryDto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Freelance Platform API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface ProjectCategoryDto
 */
export interface ProjectCategoryDto {
    /**
     * 
     * @type {number}
     * @memberof ProjectCategoryDto
     */
    projectCategoryId?: number;
    /**
     * 
     * @type {string}
     * @memberof ProjectCategoryDto
     */
    name?: string;
    /**
     * 
     * @type {boolean}
     * @memberof ProjectCategoryDto
     */
    isVisible?: boolean;
    /**
     * 
     * @type {Date}
     * @memberof ProjectCategoryDto
     */
    createdAt?: Date;
    /**
     * 
     * @type {Date}
     * @memberof ProjectCategoryDto
     */
    updatedAt?: Date;
}

/**
 * Check if a given object implements the ProjectCategoryDto interface.
 */
export function instanceOfProjectCategoryDto(value: object): value is ProjectCategoryDto {
    return true;
}

export function ProjectCategoryDtoFromJSON(json: any): ProjectCategoryDto {
    return ProjectCategoryDtoFromJSONTyped(json, false);
}

export function ProjectCategoryDtoFromJSONTyped(json: any, ignoreDiscriminator: boolean): ProjectCategoryDto {
    if (json == null) {
        return json;
    }
    return {
        
        'projectCategoryId': json['projectCategoryId'] == null ? undefined : json['projectCategoryId'],
        'name': json['name'] == null ? undefined : json['name'],
        'isVisible': json['isVisible'] == null ? undefined : json['isVisible'],
        'createdAt': json['createdAt'] == null ? undefined : (new Date(json['createdAt'])),
        'updatedAt': json['updatedAt'] == null ? undefined : (new Date(json['updatedAt'])),
    };
}

export function ProjectCategoryDtoToJSON(json: any): ProjectCategoryDto {
    return ProjectCategoryDtoToJSONTyped(json, false);
}

export function ProjectCategoryDtoToJSONTyped(value?: ProjectCategoryDto | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'projectCategoryId': value['projectCategoryId'],
        'name': value['name'],
        'isVisible': value['isVisible'],
        'createdAt': value['createdAt'] == null ? undefined : ((value['createdAt']).toISOString()),
        'updatedAt': value['updatedAt'] == null ? undefined : ((value['updatedAt']).toISOString()),
    };
}